requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
Pillow
python-dotenv
django-filter